Disable portrait orientations, require full screen
This disables portrait orientation on iPhone and disables the portrait and upside down orientations on iPad. It also makes the app require full screen on iPad, which disables Slide Over and Split View. The game does not look right in these situations and can even cause the game to freak out and lock up.